# service/aspnetcore/src/thumbnailService/Controllers/JobProcessingApi.cs
/*
 * 3D Model Thumbnail Service
 *
 * Service for generating thumbnails from 3D models
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 * Generated by: https://openapi-generator.tech
 */

using System;
using System.Collections.Generic;
using System.ComponentModel.DataAnnotations;
using Microsoft.AspNetCore.Authorization;
using Microsoft.AspNetCore.Mvc;
using Microsoft.AspNetCore.Http;
using Swashbuckle.AspNetCore.Annotations;
using Swashbuckle.AspNetCore.SwaggerGen;
using Newtonsoft.Json;
using thumbnailService.Attributes;
using thumbnailService.Models;

namespace thumbnailService.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public abstract class JobProcessingApiController : ControllerBase
    { 
        /// <summary>
        /// 
        /// </summary>
        /// <param name="id"></param>
        /// <response code="200">Information about the job requested.</response>
        [HttpGet]
        [Route("/queryJob")]
        [ValidateModelState]
        [SwaggerOperation("GetQueryJob")]
        [SwaggerResponse(statusCode: 200, type: typeof(JobFullInformation), description: "Information about the job requested.")]
        public abstract IActionResult GetQueryJob([FromRoute (Name = "id")][Required]Guid id);

        /// <summary>
        /// 
        /// </summary>
        /// <param name="id"></param>
        /// <response code="200">Returns the thumbnail of the posted model.</response>
        /// <response code="201">Model thumbnail now processing.</response>
        [HttpGet]
        [Route("/startJob")]
        [ValidateModelState]
        [SwaggerOperation("GetStartJob")]
        [SwaggerResponse(statusCode: 200, type: typeof(IconResponse), description: "Returns the thumbnail of the posted model.")]
        [SwaggerResponse(statusCode: 201, type: typeof(IconResponse), description: "Model thumbnail now processing.")]
        public abstract IActionResult GetStartJob([FromRoute (Name = "id")][Required]Guid id);

        /// <summary>
        /// 
        /// </summary>
        /// <param name="id"></param>
        /// <response code="200">Returns the thumbnail of the posted model.</response>
        [HttpGet]
        [Route("/getThumbnail")]
        [ValidateModelState]
        [SwaggerOperation("GetThumbnail")]
        [SwaggerResponse(statusCode: 200, type: typeof(IconResponse), description: "Returns the thumbnail of the posted model.")]
        public abstract IActionResult GetThumbnail([FromRoute (Name = "id")][Required]Guid id);

        /// <summary>
        /// 
        /// </summary>
        /// <param name="jobDescription">Creation of job</param>
        /// <response code="200">Returns the thumbnail of the posted model.</response>
        [HttpPost]
        [Route("/createJob")]
        [Consumes("application/json")]
        [ValidateModelState]
        [SwaggerOperation("PostCreateJob")]
        [SwaggerResponse(statusCode: 200, type: typeof(JobInformation), description: "Returns the thumbnail of the posted model.")]
        public abstract IActionResult PostCreateJob([FromBody]JobDescription jobDescription);

        /// <summary>
        /// 
        /// </summary>
        /// <param name="modelID">The job Guid of the model to submit</param>
        /// <param name="inlineObject"></param>
        /// <response code="200">Ok</response>
        [HttpPost]
        [Route("/submitModel")]
        [Consumes("multi/form-data")]
        [ValidateModelState]
        [SwaggerOperation("PostSubmitModel")]
        public abstract IActionResult PostSubmitModel([FromQuery (Name = "modelID")][Required()]Guid modelID, [FromBody]InlineObject inlineObject);
    }
}